+++ examples/cross_val/scripts/cross_val_algo_random_forest.py
@@ -22,7 +22,7 @@ current_directory = os.path.dirname(__file__)
 assets_keys_path = os.path.join(current_directory, "../../titanic/assets_keys.json")
 folds_keys_path = os.path.join(current_directory, "../folds_keys.json")
 
-client = substra.Client.from_config_file(profile_name="node-1")
+client = substra.Client.from_config_file(profile_name="organization-1")
 
 print(f"Loading existing asset keys from {os.path.abspath(assets_keys_path)}...")
 with open(assets_keys_path, "r") as f:
@@ -69,5 +69,5 @@ with open(folds_keys_path, "w") as f:
 
 print(f"Tuples keys have been saved to {os.path.abspath(folds_keys_path)}")
 print("\nRun the following commands to track the status of the tuples:")
-print(f'    substra list traintuple -f "traintuple:tag:{tag}" --profile node-1')
-print(f'    substra list testtuple -f "testtuple:tag:{tag}" --profile node-1')
+print(f'    substra list traintuple -f "traintuple:tag:{tag}" --profile organization-1')
+print(f'    substra list testtuple -f "testtuple:tag:{tag}" --profile organization-1')