+++ examples/titanic/scripts/add_dataset_metric.py
@@ -46,7 +46,7 @@ def progress_bar(length):
 current_directory = os.path.dirname(__file__)
 assets_directory = os.path.join(current_directory, "../assets")
 
-client = substra.Client.from_config_file(profile_name="node-1")
+client = substra.Client.from_config_file(profile_name="organization-1")
 
 DATASET = {
     "name": "Titanic",