+++ examples/titanic/scripts/add_train_algo_random_forest.py
@@ -21,7 +21,7 @@ import substra
 current_directory = os.path.dirname(__file__)
 assets_directory = os.path.join(current_directory, "../assets")
 
-client = substra.Client.from_config_file(profile_name="node-1")
+client = substra.Client.from_config_file(profile_name="organization-1")
 
 ALGO_KEYS_JSON_FILENAME = "algo_random_forest_keys.json"
 
@@ -111,5 +111,5 @@ with open(assets_keys_path, "w") as f:
 
 print(f"Assets keys have been saved to {os.path.abspath(assets_keys_path)}")
 print("\nRun the following commands to track the status of the tuples:")
-print(f"    substra get traintuple {traintuple_key} --profile node-1")
-print(f"    substra get testtuple {testtuple_key} --profile node-1")
+print(f"    substra get traintuple {traintuple_key} --profile organization-1")
+print(f"    substra get testtuple {testtuple_key} --profile organization-1")